Define new chunked image layer
