+++ packages/core/examples/chunk_streaming/main.ts
@@ -1,6 +1,6 @@
 import {
   Idetik,
-  ImageLayer,
+  ChunkedImageLayer,
   OmeZarrImageSource,
   OrthographicCamera,
   Region,
@@ -29,9 +29,9 @@ const region: Region = [
   { dimension: "x", index: { type: "full" } },
 ];
 
-const channelProps = [{ contrastLimits: [0, 255] as [number, number] }];
+const channelProps = { contrastLimits: [0, 255] as [number, number] };
 const camera = new OrthographicCamera(left, right, top, bottom);
-const imageLayer = new ImageLayer({ source, region, channelProps });
+const imageLayer = new ChunkedImageLayer({ source, region, channelProps });
 imageLayer.debugMode = true;
 
 const overlaySelector = document.querySelector<HTMLDivElement>("#chunk-info")!;
