+++ packages/core/src/layers/image_series_layer.ts
@@ -12,6 +12,7 @@ export type ImageSeriesLayerProps = LayerOptions & {
   region: Region;
   seriesDimensionName: string;
   channelProps?: ChannelProps[];
+  lod?: number;
 };
 
 export class ImageSeriesLayer extends Layer implements ChannelsEnabled {
