+++ packages/core/src/layers/label_image_layer.ts
@@ -17,6 +17,7 @@ export type LabelImageLayerProps = LayerOptions & {
   region: Region;
   colorMap?: LabelColorMapProps;
   onPickValue?: (info: PointPickingResult) => void;
+  lod?: number;
 };
 
 export class LabelImageLayer extends Layer {
