+++ packages/core/src/layers/label_image_series_layer.ts
@@ -15,6 +15,7 @@ export type LabelImageSeriesLayerProps = LayerOptions & {
   region: Region;
   seriesDimensionName: string;
   colorMap?: LabelColorMapProps;
+  lod?: number;
 };
 
 export class LabelImageSeriesLayer extends Layer {
