+++ packages/core/src/utilities/logger.ts
@@ -19,6 +19,7 @@ type Module =
   | "ChunkManager"
   | "EventDispatcher"
   | "Idetik"
+  | "ChunkedImageLayer"
   | "ImageLayer"
   | "LabelImageLayer"
   | "WebGLRenderer"
